Remove top-level side effects from importing libdelete
If the afs python modue isn't installed, then when libdelete is
imported, this was getting displayed instead of the intended warning
about afs not being available:

>`No handlers could be found for logger "libdelete"`

This is because libdelete was attempting to import afs to check for
its availability at module scope, before the main script importing
libdelete had configured logging.

Fix this by changing `have_AFS` from a module-scope variable to a
memoized function so that the check occurs after logging is
configured.