Use uniquifying list when searching for deleted files
This change is intended to fix the following:

```
$ touch foo
$ delete foo
$ touch foo
$ expunge
Traceback (most recent call last):
  File "/usr/local/bin/expunge", line 183, in <module>
    sys.exit(main())
  File "/usr/local/bin/expunge", line 176, in main
    rv += expunge(deleted_files, options)
  File "/usr/local/bin/expunge", line 90, in expunge
    os.unlink(f)
OSError: [Errno 2] No such file or directory: './.#foo'
$
```

The problem was that "./.#foo" was getting added to the deleted file
list twice, because it glob-matched twice, once in its undeleted form
and again in its deleted form.

To avoid this issue we use a uniquifying list class so that we don't
add the same path to our list of deleted files twice.

We use a uniquifying list rather than a set to preserve order.